MAINT: Remove old hooks file
