+++ src/contexts/SelectedSpotsProvider.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useList } from 'react-use'
 import { ListActions } from 'react-use/lib/useList'
 
-import { SpotDTO } from 'hooks/usePlanEvents'
+import { SpotDTO } from './CurrentPlanProvider'
 
 export const SelectedSpotsContext = React.createContext<Array<SpotDTO>>([])
 export const SetSelectedSpotsContext =
