+++ src/hooks/useEventFactory.ts
@@ -2,7 +2,11 @@ import * as React from 'react'
 import { EventInput } from '@fullcalendar/react'
 import { v4 as uuidv4 } from 'uuid'
 
-import { MoveEvent, ScheduleEvent, SpotEvent } from 'hooks/usePlanEvents'
+import {
+  MoveEvent,
+  ScheduleEvent,
+  SpotEvent,
+} from 'contexts/CurrentPlanProvider'
 
 type BuildMoveParams = {
   start: MoveEvent['start']
